move MaximumAbsoluteDrawdownCriterion into the right package
