+++ ta4j-core/src/test/java/org/ta4j/core/criteria/MaximumAbsoluteDrawdownCriterionTest.java
@@ -30,7 +30,6 @@ import org.ta4j.core.BaseTradingRecord;
 import org.ta4j.core.Position;
 import static org.ta4j.core.TestUtils.assertNumEquals;
 import org.ta4j.core.Trade;
-import org.ta4j.core.criteria.pnl.MaximumAbsoluteDrawdownCriterion;
 import org.ta4j.core.mocks.MockBarSeriesBuilder;
 import org.ta4j.core.num.NumFactory;
